Alterado idUsuario para Long

+++ src/main/java/com/estoqueapi/EstoqueApi/Entidades/Usuarios.java
@@ -1,7 +1,6 @@
 package com.estoqueapi.EstoqueApi.Entidades;
 
 import com.estoqueapi.EstoqueApi.Enums.PerfilUsuario;
-import org.springframework.boot.context.properties.bind.DefaultValue;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -10,7 +9,7 @@ import javax.persistence.Id;
 @Entity
 public class Usuarios {
     @Id
-    private int idUsuario;
+    private Long idUsuario;
     private String nome;
     private String senha;
 
@@ -29,11 +28,11 @@ public class Usuarios {
         this.email = email;
     }
 
-    public int getIdUsuario() {
+    public long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(int idUsuario) {
+    public void setIdUsuario(Long idUsuario) {
         this.idUsuario = idUsuario;
     }
